+++ tools/kystrap/main.go
@@ -3,7 +3,7 @@ Copyright © 2023 Raphael Thurnherr <rapha@kyve.network>
 */
 package main
 
-import "github.com/KYVENetwork/kyvejs/tools/kystrapper/cmd"
+import "github.com/KYVENetwork/kyvejs/tools/kystrap/cmd"
 
 func main() {
 	cmd.Execute()
