+++ tools/kystrap/template-strings.yml
@@ -1,5 +1,5 @@
 ---
-name: "<undefined>"
+name: "<set-by-cmd-tool>"
 lang:
   go:
       path: "templates/go"
